Change sizes for ext4 images
Change page_size to 512 and extra_size to 0. This is needed to use ext4
filesystem, which is used beyond apiLevel 18, on goldfish.  A new
hardware property has been added to specify which filesystem to use:
ext4 or yaffs2. The default for this property is to use yaffs2 if
apiLevel <= 18 and use ext4 otherwise. It can be overridden in AVD's
config.ini by specifying "hw.useext4=yes".

Bug: http://b.android.com/38561
Change-Id: I978afe0f742795beef66e7eaee59a8d9e76d4602