Auto-detect ext4 image file.
Although the existing code can detect the usage of ext4 file
system by the API level.  However, there might be some
problem when we are building and running the emulator in
the source tree.

This CL workaround this issue by checking the magic word in
the ext4 header.  If the system.img has the magic word at
0x438, then add hw.useext4 to the hardware-qemu.ini.

Related Issue:
http://code.google.com/p/android/issues/detail?id=38561

Change-Id: I0b3606911cba5f52d92157accaaba3b2caef85c1
Signed-off-by: Ricardo Salveti de Araujo <ricardo.salveti@canonical.com>