Enable sending NITZ time, and fix its format
Sending NITZ (network time) from the simulated modem was
flagged out; now it is sent with each signal strength update
(emulator console command 'gsm signal').

Also, the format of the time string was not according to the
RIL API spec. Note that this needs a corresponding change
in the reference RIL.

Change-Id: I731676f970a8650b4c49063a31816723d85d0387
Signed-off-by: Martti Piirainen <martti.piirainen@canonical.com>