Fix touch screen emulation by removing mouse emulation
The emulated input device in qemu is a hybrid of a keyboard, relative
mouse and a touchscreen. Since libinput allows tracking mouse devices
with absolute positions the evdev axes of the different modes overlap.
Additionally with this mouse device active on boot, ubuntu-touch would
always run in windowed mode.
To resolve that the relative events and mouse event have been
disabled, now the device will always act as a multitouch touchscreen.

Change-Id: Id01d2a4afd5673e9f21354d658f5becf47be2eaf
Signed-off-by: Andreas Pokorny <andreas.pokorny@canonical.com>